+++ app/Http/Controllers/RoleController.php
@@ -19,6 +19,8 @@ class RoleController extends Controller
 
     public function index(): View
     {
+        $this->authorize('role.index');
+
         $roles = $this->rolesRepo->getAllRoles();
 
         return view('role.index', compact('roles'));
@@ -26,11 +28,15 @@ class RoleController extends Controller
 
     public function create(): View
     {
+        $this->authorize('role.create');
+
         return view('role.create');
     }
 
     public function store(RoleStoreRequest $request): RedirectResponse
     {
+        $this->authorize('role.create');
+
         $this->rolesRepo->storeRole($request);
 
         return redirect()->route('role.index')
@@ -39,16 +45,22 @@ class RoleController extends Controller
 
     public function show(Role $role): View
     {
+        $this->authorize('role.show');
+
         return view('role.view', compact('role'));
     }
 
     public function edit(Role $role): View
     {
+        $this->authorize('role.update');
+
         return view('role.edit', compact('role'));
     }
 
     public function update(UpdateRoleRequest $request, Role $role): RedirectResponse
     {
+        $this->authorize('role.update');
+
         $this->rolesRepo->updateRole($request, $role);
 
         return redirect()->route('role.index')
@@ -57,6 +69,8 @@ class RoleController extends Controller
 
     public function destroy(Role $role): RedirectResponse
     {
+        $this->authorize('role.disable');
+
         $this->rolesRepo->delete($role);
 
         return redirect()->route('role.index')
